Fix Funk parser for new URL
The JSON with all channel videos is now located at a different URL.

For end users this means they have to adapt the config files.
The numeric channel id is now enough.
F.e. "walulis-daily-12068" has to be changed to "12068"
(see the example in the test cases).

Furthermore, the datetime format inside the JSON changed.